some major refactoring (needs cleaning)

+++ src/transforms.py
@@ -8,7 +8,7 @@ class TargetNormalize(object):
 
     def __call__(self, value):
         value = (value - self.mean) / self.std
-        return value.astype("float32")
+        return value
 
     def revert(self, value):
         value = value * self.std + self.mean
